feat(sqlite): add environment option
Added the 'prismaEnvVarName' environment option to override the environment
variable name used for Prisma connections which is necessary for situations
where the variable is not the standard "DATABASE_URL" default. As an example
when using PlanteScale as a database provider the connection variable needs
to be "PLANETSCALE_PRISMA_DATABASE_URL".

Updated the README to document the new environment option as well as make
some minor edits related to SQLite support.